+++ index.html
@@ -31,8 +31,6 @@
               type="text"
               placeholder="Search for a movie"
               class="search-box-input"
-              oninput="search()"
-              onkeypress="pressEnter(event)"
               autocomplete="off"
             />
             <div id="similarTitle"></div>
@@ -49,6 +47,6 @@
         <div id="cardsDiv" class="movie-cards-div"></div>
       </div>
     </div>
-    <script src="./js/home.js"></script>
+    <script type="module" src="./js/home.js"></script>
   </body>
 </html>
